net: ti: icssg-prueth: move phy_connect()/phy_disconnect()
commit 91ea70eb3334d9d5b2c9ea13b274bbeb5bc678f9 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The phy_connect() and phy_disconnect() functions are currently called
when the interface is brought up and down using ifconfig respectively.
This causes an issue when ethtool is used after bringing down the
interface. This issue was reported by Siemens.

Move the phy_connect()/phy_disconnect() logic to probe and remove
respectively to solve this bug.

Signed-off-by: Puranjay Mohan <p-mohan@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Link: https://serenity.dal.design.ti.com/lore/linux-patch-review/20220509053931.4552-1-p-mohan@ti.com
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>