fpga: Add ps-pl reset handling through xilinx afi driver
commit e5e6e431f6749e4e167b2b367acbe85c87cf9540 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Ideally, the PL IPs must be held in reset during the clock switch over
but in the current flow, we are doing the ps-pl reset before configuring
the clocks and there is no reset handling after clock change. So we can’t
guarantee the functionality if we ignore reset during clock switch over.

To address this issue this patch adopts the reset handling support for
ps-pl resets after configuring the PS_PL clocks and AFI configurations.

Signed-off-by: Nava kishore Manne <nava.manne@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>