fpga: Add ps-pl reset handling through xilinx afi driver
commit 7542eb398b6713a6941ed48f93012bf161e91521 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Ideally, the PL IPs must be held in reset during the clock switch over
but in the current flow, we are doing the ps-pl reset before configuring
the clocks and there is no reset handling after clock change. So we can’t
guarantee the functionality if we ignore reset during clock switch over.

To address this issue this patch adopts the reset handling support for
ps-pl resets after configuring the PS_PL clocks and AFI configurations.

Signed-off-by: Nava kishore Manne <nava.manne@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>