soc: octeontx2-sdp: Fix SDP output backpressure
commit 67291888a8685cbedc00ad26b226d42d9db5acf5 from
git@git.assembla.com:cavium/WindRiver.linux.git

Fix Octeon SDP output backpressure.
Note that backpressure only works when a single SDP queue
per VF or PF is used.  This is due to how SDP queues
interract with NIX output TL4 queues, and the requirement
that there must be a 1-1 TL4 to SDP queue mapping for
backpressure to work.  In order for backpressure to
work with more SDP queues, NIX queue allocation
for SDP must be changed to match.

Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I0b251e1303a70fbb57aa21c0764e5c45771f501f
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/76850
Reviewed-by: Radha Chintakuntla <radhac@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>