soc: imx: gpcv2: Drop the unneeded imx_pgc_power_down()
The imx_pgc_power_down() sneaked in due to a mistake
in commit 019cfa5ce68d ("LF-4674 soc: imx: gpcv2: Fix
imx7d pcie & csi failure caused by power domain"),
but there is none of usage of it. So delete it to fix
the "defined but not used" build error.

Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>