xilinx-ams: Fix UBSan Warning
commit 74f12f0b1fd0e10137609327f16900a954d68378 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

The issue was that a 64 bit datatype was being left shifted by
more than 63 bits due to the scan index of the AMS control
channels.

There is no functional need for these control channels to be
included in the scan_mask. This leads to the fix being a
range check on the scan index before shifting.

Addresses-UBSAN: shift-out-of-bounds
Signed-off-by: Conall O'Griofa <conall.o'griofa@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>