misc: bphy: prevent out-of-bound array iteration
commit 261b4ac996eb1f90370ac97ad30d9335fae8b2b9 from
git@git.assembla.com:cavium/WindRiver.linux.git

Interrupt cleanup procedure is executed regardless on
driver state. On non-BPHY boards or with firmware not
supporting runtime interrupt number detection this
results in use of erroneous values as maximum array
index. With this patch the situation is detected and
release procedure for non-initialized irqs is skipped.

Signed-off-by: Jakub Palider <jpalider@marvell.com>
Change-Id: Id7bdbe56213ffec0e058897ac4a1fccd680de7c4
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/76964
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>