serial: core: Do stop_rx in suspend path for console if console_suspend is disabled
commit 1485291493cc7b627187e916d4612c9641f7e5fa from
https://github.com/nxp-auto-linux/linux

For the case of console_suspend disabled, if back to back suspend/resume
test is executed, at the end of test, sometimes console would appear to
be frozen not responding to input. This would happen because, during
resume, rx transactions can come in before system is ready, malfunction
of rx happens in turn resulting in console appearing to be stuck.

Do a stop_rx in suspend sequence to prevent this.

This is a cherry-pick of:
commit c9d2325cdb92 ("serial: core: Do stop_rx in suspend path for
		      console if console_suspend is disabled")

Issue: ALB-8958
Issue: ALB-9101
Signed-off-by: Vijaya Krishna Nivarthi <quic_vnivarth@quicinc.com>
Link: https://lore.kernel.org/r/1652692810-31148-1-git-send-email-quic_vnivarth@quicinc.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>