HSD #1509151525: fpga: altera-cvp: Truncated bitstream error support
commit 246dd469d5c614849feebd967768e7c235c0d8e5 from
https://github.com/altera-opensource/linux-socfpga.git socfpga-6.6.22-lts.

To support the error handling of a truncated bitstream sent.
The current AIB CvP firmware is not capable of handling a
data stream smaller than 4096bytes. The firmware's limitation
causes a hung-up as it's DMA engine waits forever for the
completion of the instructed 4096bytes.
To resolve this design limitation, both firmware and CvP
driver made several changes. At the CvP driver, we just
have to ensure that anything lesser than 4096bytes are
padded with extra bytes. The CvP will then, initiate the
tear-down by clearing the START_XFER and CVP_CONFIG bits.
We should also check for CVP_ERROR during the CvP completion.
A send_buf which is always 4096bytes is used to copy the
data during every transaction.

Signed-off-by: Ang Tien Sung <tien.sung.ang@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>