vc04_services: bcm2835-codec: Set crop_height for compressed formats
commit  0cd33a0bba43fb15d83a44ae19ccecf656aa1070 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

In particular for the encoder where the CAPTURE format dictates
the parameters given to the codec we need to be able to set the
value passed as the crop_height for the compressed format.
There's no crop available for cropped modes, so always set
crop_height to the requested height.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>