vc04_services: bcm2835-codec: Set port format from s_selection
commit  6b945e6f05d5cb0a299dbbb9bdd285aff330a204 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

s_selection allows the crop region of an uncompressed pixel
format to be specified, but it wasn't passing the setting on to
the firmware. Depending on call order this would potentially
mean that the crop wasn't actioned.

Set the port format on s_selection if we have a component created.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>