dpaa2-switch: cleanup the egress flood of an unused FDB
commit  ebcc64da04bb5d1a3b2b14abea201913240860e6 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

In case a switch interface is joining a bond device or a bridge one, its
FDB might change. In case this happens, we have to recreate the egress
flood setup of the FDB that we just left.
For this to happen, keep track of the old FDB and just call
dpaa2_switch_fdb_set_egress_flood() on it.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>