dpaa2-switch: cleanup the egress flood of an unused FDB
commit 54710002fc7ef03f5ef18b54c4347bf7338aa502 from
https://github.com/nxp-imx/linux-imx.git

In case a switch interface is joining a bond device or a bridge one, its
FDB might change. In case this happens, we have to recreate the egress
flood setup of the FDB that we just left.
For this to happen, keep track of the old FDB and just call
dpaa2_switch_fdb_set_egress_flood() on it.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>