usb: dwc3: xilinx: Fix error handling when getting USB3 PHY
commit a434c97c4c3024d01ef46640b8965bf53a2fc389 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

The code that looked up the USB3 PHY was ignoring all errors other than
EPROBE_DEFER in an attempt to handle the PHY not being present. Fix and
simplify the code by using devm_phy_optional_get and dev_err_probe so
that a missing PHY is not treated as an error and unexpected errors are
handled properly.

When no USB3 PHY 'usb3-phy' property is specified in the device tree,
then zynqmp board work as a USB2.0 mode only.

Signed-off-by: Piyush Mehta <piyush.mehta@xilinx.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>