usb: dwc3: xilinx: fix usb3 non-wakeup source resume failure
commit ddc809972664fc805b446dc0e85104179c5233d9 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

When USB is in super-speed mode and disabled as a wakeup source,
observed that on the resume path, lanes have not been configured
properly in the phy-zynqmp driver.

To resolved above issue, added phy_init on resume and phy_exit on
suspend path to configure the GT lanes correctly.
The re-initialization of phy, reset the device and re-enumerate
the USB subsystem.

This use-case is specific to zynqmp and usb super-speed mode.

Note:
With the above fix below warnings messages are coming, but USB
functionality is not impacting.
Disable of device-initiated U1 failed
Disable of device-initiated U2 failed

Signed-off-by: Piyush Mehta <piyush.mehta@xilinx.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>