octeontx2-pf: qos: use quantum value 0x7fff_ffff to identify static priority child
commit f65e4f2a302303c7c0f35a08cd1aad9cc231f0d3 from
git@git.assembla.com:cavium/WindRiver.linux.git

The current HTB offload driver code assumes that when user doesn't
provide a quantum value (=0), that htb class is treated as a static
priority child. However the kernel htb scheduler code calculates the
quantum value when quantum=0, but the htb-offload is called before
re-calculating the quantum. This part of code needs to be fixed in
the Linux kernel htb scheduler code for htb-offload.

The side-effect of this piece of code is, when the htb class is dumped
by the user, the htb class parameters display the calculated quantum
value instead of zero.

To change this behaviour, the code is modified such that the the quantum
value 0x7fff_ffff (max signed integer) is used to identify static priority
child. This achieves symmetric set/get value of quantum value from user
space. The quantum value zero is not valid.

Change-Id: I2df6ec2476c51d550dd6a6e59940529ffc658456
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/78358
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>