LF-6273-1 arm64: dts: imx93: Add 'fsl,gpr' property for lcdif node
commit eb52e5ec46c574f9e2cfd381631ffea6062da7b8 from
https://github.com/nxp-imx/linux-imx.git

As a temporary solution, we need to get the media_blk_ctrl phandle
through the added 'fsl,gpr' lcdif node property to control lcdif
QoS and cache in the lcdif driver to achieve better system performance
and lcdif memory access performance.  In the future, we'll use proper
and generic APIs in the lcdif driver to do the control instead of
accessing the mediamix block control register by our own.

Cc: Sandor Yu <Sandor.yu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>