drm: xlnx: zynqmp_dp: Fix kernel dump messages when the DP cable is disconnected
commit 88b8a79111e6c2e58c2b498372f809d0682b1b6c from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Below error message is shown when the DP cable is disconnected.

[  108.780458] zynqmp-display fd4a0000.display: failed to write a byte to the DPCD: -19

This error message will confuse the user because the aux channel is not
present after unplugging the cable, so any read / write operation to the
DPCD register will fail.

So adding a drm_dp_dpcd_readb() before we write to the DPCD register. And
if the read fails, skip dpcd_writeb and continue in the function to power
down phy and disable the audio to avoid a false alarm message.

Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>
Reviewed-by: Portia Stephens <portia.stephens@canonical.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>