net: phylink: handle NA interface mode in phylink_fwnode_phy_connect()
commit e865e3202da5885051e66de31e5afcd2068140fa from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit a18e6521a7d95dae8c65b5b0ef6bbe624fbe808c upstream.

Commit 4904b6ea1f9db ("net: phy: phylink: Use PHY device interface if
N/A") introduced handling for the phy interface mode where this is not
known at phylink creation time. This was never added to the OF/fwnode
paths, but is necessary when the phy is present in DT, but the phy-mode
is not specified.

Add this handling.

Signed-off-by: Russell King (Oracle) <rmk+kernel@armlinux.org.uk>
Acked-by: Florian Fainelli <f.fainelli@gmail.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>