phy: ti: gmii-sel: Add support for CPSW9G GMII SEL in J721E
commit bb6200e31b0e2c8d70f403e63d0bc19ec2672be4 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Each of the CPSW9G ports in J721E support additional modes like QSGMII.
Add a new compatible for J721E to support the additional modes.

In TI's J721E, each of the CPSW9G ethernet interfaces can act as a
QSGMII or QSGMII-SUB port. The QSGMII interface is responsible for
performing auto-negotiation between the MAC and the PHY while the rest of
the interfaces are designated as QSGMII-SUB interfaces, indicating that
they will not be taking part in the auto-negotiation process.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>