LF-6937-15 clk: imx: blk-ctrl: update mediamix/hdmimix compatible string
commit eb9120acb1880daf41232434c5625e31d60379cb from
https://github.com/nxp-imx/linux-imx

Use a new compatible string for clk blk-ctrl.
Since we are going to support power domain blk-ctrl driver
which use same compatbile string with the current clk blk-ctrl
driver. That means two driver shares same compatible string which would
cause issue. This patch is not good, but since we will drop the clk
blk-ctrl in the end, so ..!

Reviewed-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>