octeontx2-bphy-netdev: cnf95n: ptp: Fix ptp clock counter read
commit e6936257ec2ccc59f51d3189e409f531ee75c4e8 from
git@git.assembla.com:cavium/WindRiver.linux.git

The PTP input clock frequency is passed to the driver using an ioctl
OTX2_RFOE_IOCTL_PTP_OFFSET and driver adds ptp-to-bcn offset to ptp
clock time. This patch adds ptp-to-bcn offset to ptp clock in
cc->read() callback. The details of whether to add BCN offset when
ptp is in slave mode is unknown at this point of time.

Change-Id: I2c3dbea8765a0660448f51fe692f2bf78c904190
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/79971
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>