crypto: sa2ul - Check engine status before enabling
commit 59d0b5f47f2ab4c15f6e9393b8b12f70673530f7 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

There is a engine status register that can be used to check if the
different HW crypto engines are enabled. Check that first and then only
try to enable the engines if they are not already on.

This has a couple benefits. First we don't need to use match_data for
this. Second, this driver can now work on HS devices where the engine
control registers are read-only and writing causes a firewall exception.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>