ARM: zynq: Deprecate zynq ocmc driver
commit d379c8482478866d5722e626a222cc48ee637730 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Zynq OCM driver was never upstreamed and in general, it is a little bit
smarter than the mmio-sram driver. There is code which reads OCM
configuration (low/high) and deals with it. The truth is that almost all
boards are using OCM at the high address where the mmio-sram driver should
be enough. So remove the zynq-com driver and use mmio-sram if someone wants
to map OCM.

Signed-off-by: Appana Durga Kedareswara rao <appana.durga.kedareswara.rao@amd.com>
Signed-off-by: Michal Simek <michal.simek@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>