fpga: zynqmp: Make word align the configuration data
commit a2f0866911dd9e50828f2ed752525211c4c3a437 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

To avoid the unwanted copies at firmware(PMUFW) this patch provides
a fix to align PL configuration data if the data is not word-aligned.
It improves the PL configuration timings to process unaligned bit
files.

To align the configuration data this patch adds a few padding bytes and
these additional padding bytes will not create any functional impact on
the PL configuration.

Signed-off-by: Nava kishore Manne <nava.kishore.manne@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>