dtoverlays: Use PDAF variant for imx258 overlay
commit  aabca34878d5c4493c8ec984f16abea8e5f31845 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

There are a number of variants of the imx258 sensor. The only
version that is currently made into a module for use with the Pi
is from Vision Components, and that supports PDAF and needs the
appropriate register configuration.

Update the compatible string accordingly.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>