media: i2c: ov9281: Correct min/def vts for 640x400
commit  5b81100f0e3680bdbea9caac065630335af5bbff from
https://github.com/raspberrypi/linux.git rpi-5.15.y

A VTS of 421 in the 640x400 mode only streams at 130.2fps
instead of the expected ~261fps. This appears to be an invalid
value for the sensor, as a VTS of 422 works fine to give 259.79fps.

Set the minimum (and default) to 422.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>