octeontx2-pf: qos: Fix dwrr child priority
commit 21568f0298f82555d3eadef69d04a46c8522b4ff from
git@git.assembla.com:cavium/WindRiver.linux.git

With the existing txschq allocation technique, the dwrr children
priority is incorrectly configured by the driver. As a result, when
static and dwrr children are created, the user-programmed priorities
are not honoured. This patch fixes the issue.

Also fixed possible overflow of quantum field.

Change-Id: Icfa665ba87cb213d41b343c51c49c66731cbccaf
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/82284
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>