dtoverlays: Reshuffle image sensor overlays to allow use with muxes
commit  d2600ce4dcdaf339cd72191ac2148048c8392a75 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

There are third-party CSI-2 muxes available that ideally want to
be configured to have any of the supported image sensors.
That wants the sensor configuration to be split out into fragements
that can be included in the relevant places, and then an overall
overlay that pulls them in.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>