pinctrl: s32cc: rename "pins" property
commit 2e77d16571b2af60ac8d1dba6d5135ba5bec3421 from
https://github.com/nxp-auto-linux/linux

The "pins" property used inside the pinctrl node was specific to the
driver. Other drivers don't use this property. It specifies the range
of MSCR/IMCR registers that are valid and correspond to the "reg"
property.

Change it into "nxp,pins". Also, remove the phandle reference from it
since it is always referencing the parent pinctrl node.

Issue: ALB-9022
Issue: ALB-9098
Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>