pinctrl: s32cc: split memory region for S32G SIUL21 MSCRs
commit 739a16260cbe222cac2849a22d1e078f0384ee45 from
https://github.com/nxp-auto-linux/linux

There is a gap for the SIUL21 module for S32G SoCs. GPIOs
123-143 are not available. Currently, trying to export a GPIO in that
region would fail because the corresponding `pinctrl_pin_desc` cannot
be found. However, the memory region is still reserved.

The problem would be seen when trying to use the same DTB in U-Boot
where the pin descriptions are not hardcoded.

Split the current memory region for S32G's SIUL21 MSCRs. Update the
available `pins` range for the pinctrl driver.

Issue: ALB-9022
Issue: ALB-9098
Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>