Input: Add tps65219 interrupt driven powerbutton
commit 6ffa48a450f631e5247f5f43d14686bc6aa5b4eb from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

TPS65219 has different interrupts compared to other TPS6521* chips.
TPS65219 defines two interrupts for the powerbutton one for push and one
for release.

This driver is very simple in that it maps the push interrupt to a key
input and the release interrupt to a key release.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Signed-off-by: Jerome Neanne <jneanne@baylibre.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>