v4l: xilinx: sdirxss: No payload handling for SDI 3G Mode
commit 80dea87a5755f07bcce9c09d2a24feee9322c6d1 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

This patch handles the video locking for SDI stream in 3G mode, even if the
SDI stream doesn't contain ST352 payloads. This is handled by controlling
the VPID bit in Module ctrl register and updating the v4l2 framework with
proper format details based on the stream information.

Signed-off-by: Ramanjaneyulu Uppari <ramanjaneyulu.uppari@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>