v4l: xilinx: sdirxss: No payload handling for SDI 3G Mode
commit 7a99325593decc0491e7770f0392a9e02d970376 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

This patch handles the video locking for SDI stream in 3G mode, even if the
SDI stream doesn't contain ST352 payloads. This is handled by controlling
the VPID bit in Module ctrl register and updating the v4l2 framework with
proper format details based on the stream information.

Signed-off-by: Ramanjaneyulu Uppari <ramanjaneyulu.uppari@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>