soc: xilinx: fix unhandled SGI warning message
commit 60b013e2c371ea60e55b3b570673e418eba54746 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Xen broadcasts SGI to each VM when multiple VMs run on Xen hypervisor. In
such case spurious SGI is expected if one event is registered by one VM and
not registered by another VM. We let users know that Unhandled SGI is not
error and expected if kernel is running on Xen hypervisor.

Signed-off-by: Tanmay Shah <tanmay.shah@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>