irqchip/gic-v3: Detect LPI invalidation MMIO registers
commit 4c09e8b9f9328b331da9b632fdcd4d9081523b64 from
git@git.assembla.com:cavium/WindRiver.linux.git

Since GICv4.1, an implementation can offer the same MMIO-based
implementation as DirectLPI, only with an ITS. Given that this
can be hugely beneficial for workloads that are very LPI masking
heavy (although these workloads are admitedly a bit odd).

Interestingly, this is independent of RVPEI, which only *implies*
the functionnality.

So let's detect whether the implementation has GICR_CTLR.IR set,
and propagate this as DirectLPI to the ITS driver.

While we're at it, repaint the GICv3 banner so that we advertise
the various capabilities at boot time to be slightly less invasive.

Signed-off-by: Marc Zyngier <maz@kernel.org>
Link: https://lore.kernel.org/r/20220405183857.205960-3-maz@kernel.org

Backported to Linux 5.4
Signed-off-by: Linu Cherian <lcherian@marvell.com>

Change-Id: Ia8240a7a8e0a769f5d6af026cb9f16de3f8c7c72
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/82850
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>