media: bcm2835-unicam: Correctly handle FS + FE ISR condtion
commit  e9b05bbd7379ef62cb32e67e33b96dd5c753eebb from
https://github.com/raspberrypi/linux.git rpi-6.1.y

If we get a simultaneous FS + FE interrupt for the same frame, it cannot be
marked as completed and returned to userland as the framebuffer will be refilled
by Unicam on the next sensor frame. Additionally, the timestamp will be set to 0
as the FS interrupt handling code will not have run yet.

To avoid these problems, the frame is considered dropped in the FE handler,
and will be returned to userland on the subsequent sensor frame.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>