dt-bindings: usb: dwc3: Add snps,xhci-reset-on-resume quirk
commit 649502579cdd2afcf77e27f60f8ff0c01bad459c from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Some controllers might lose power during suspend, so they are requires
re-initialization. During the suspend cycle when USB is not pluh-in or
USB not set as a wakeup source then power management turn off the clock
and power, to save power.

To address above suspend-resume use-case snps,xhci-reset-on-resume quirk
is added, which will reset the host controller on resume.

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>