arm64: dts: ti: k3-am625: Introduce operating-points table
commit df67dd915b970577ee60d072fce5d96d54f4b816 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Introduce an operating-points table for the A53 cores, containing only
frequency values as this platform operates on a fixed voltage for the
CPUs. Also provide opp-supported-hw values to ensure appropriate OPPs
are enabled based on which type of silicon is in use.

The latency between pre and post frequency transition was measured in
CPUFreq driver for all combinations of OPP changes. The average value
was selected as overall clock-latency-ns.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>