HSD #18022323310: spi: cadence-quadspi: Disable irqs during indirect reads
commit cd2ced4a17280d0d84c4af69eb225cb3ca5a534f from
https://github.com/altera-opensource/linux-socfpga.git socfpga-6.6.22-lts.

On architecture where reading the SRAM is slower than the pace at
controller fills it, with interrupt enabled while reading from
SRAM FIFO causes unwanted interrupt storm to CPU.

The inner "bytes to read" loop never exits and waits for the completion
so it is enough to only enable the watermark interrupt when we
are out of bytes to read, which only happens when we start the
transfer (waiting for the FIFO to fill up initially) if the SRAM
is slow.

So only using read watermark interrupt, as the current implementation
doesn't utilize the SRAM full and indirect complete read interrupt.
And disable all the read interrupts while reading from SRAM.

Signed-off-by: Niravkumar L Rabara <niravkumar.l.rabara@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>