s32g274a: ddr: Fixed read_lpddr4_mr function
commit 742ada1961dc504d190fced86f5e6599b95b5021 from
https://github.com/nxp-auto-linux/linux

The workaround for this errata should read only RANK 0
and check both dies.

Currently, the Linux driver only cleared rank0 bit, and reads MR from
RANK 1 which is not an issue on dual rank devices such as S32G274A-EVB,
but it is not correct on single rank devices, which do not have RANK 1.

Issue: ALB-7692
Issue: ALB-9124
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>