LF-6946-2 drm/imx: imx8qxp-ldb: Initialize encoder only for enabled LVDS channel
commit 46c67e9c80b98b7f6b2d450e6aed28358dd8da2e from
https://github.com/nxp-imx/linux-imx.git

We should not initialize encoders for both LVDS channels and then
cleanup encoder for disabled LVDS channel, because the encoder index
of the second LVDS channel(enabled in device tree) would be reused
by an upcoming encoder if the first LVDS channel is disabled in device
tree. So, ensure LVDS channel is enabled in device tree first, and then
initialize encoder for it.

Cc: Sandor Yu <Sandor.yu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>