LF-6946-3 drm/imx: imx8qm-ldb: Initialize encoder only for enabled LVDS channel
commit 689128cc50edebc9a5d342929d94fef701f18a4a from
https://github.com/nxp-imx/linux-imx

We should not initialize encoders for both LVDS channels and then
cleanup encoder for disabled LVDS channel, because the encoder index
of the second LVDS channel(enabled in device tree) would be reused
by an upcoming encoder if the first LVDS channel is disabled in device
tree. So, ensure LVDS channel is enabled in device tree first, and then
initialize encoder for it.

Cc: Sandor Yu <Sandor.yu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>