nvmem: siul2: Create the header file for S32CC SoC revision
commit 192e307df75045e8e0e38a6e467351e18392d2ca from
https://github.com/nxp-auto-linux/linux

The proposed API extends the S32 SIUL2 NVMEM to provide a user friendly
interface for getting the SoC revision

The consumers have to:
- create the necessary device tree binding for the NVMEM cells, e.g.
		nvmem-cells =  <&soc_revision>;
		nvmem-cell-names = "soc_revision";
- call the API passing struct s32_soc_rev parameter

The API handles the NVMEM cell get/put management as well as freeing the
allocated buffers. Deferred probe is also handled, through return value,
in case the NVME is not ready.
The API should be called in probe and its return value should be propagated
in case of error to correctly handle the deferred probing.

Issue: ALB-5926
Issue: ALB-9139
Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>