Revert "ACPI: APEI: explicit init of HEST and GHES in apci_init()"
This logically reverts 335852f7a976 ("ACPI: APEI: explicit init of HEST
and GHES in apci_init()"). The diff stat is slightly different with
the commit 335852f7a976 since there are some conflicts when that commit
was merged into linux-yocto.

On the Marvell cn96xx/cn106k SoCs, there is no presence of ACPI at all.
But it leverage some functions of ACPI for the GHES support. After the
change of 335852f7a976, the ghes_init() is invoked in the acpi_init()
instead of the initcall. Since the acpi_disabled is true for the
Marvell SoCs, the acpi_init() just bail out without running any codes
and it is also not possible to patch acpi_init() to do the ACPI
initialization on a platform with acpi_disabled. So revert this commit
and it is pretty safe to do so since the issue that commit tries to fix
doesn't exist on a acpi_disabled platform.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>