octeontx2-pf: Update PFC configuration
commit 915f2af642ce65b8d629d81e3b16caa7fb8ee91a from
git@git.assembla.com:cavium/WindRiver.linux.git

As of now we are creating/deleting Tx schedulers when user is
setting PFC on/off. The problem is if we have a running traffic on
the interface and as we are updating the sq->smq mapping on the fly,
we might loose completion interrupt for some packets. As a result of
that a watchdog reset is hit from BQL.
This patch solves the issue by simply calling interface off/on APIs
which will reconfigure all the queues. We might loss the running traffic
momentarily but that should be fine.

Fixes: 6ed58fb724bf (octeontx2-pf: Add egress PFC support)

Signed-off-by: Suman Ghosh <sumang@marvell.com>
Change-Id: I793695c09b0685783cece93c2beed9ccceabf4e7
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/84851
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/99476
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>