arm64: dts: ti: k3-j721s2: Fix the interrupt ranges property for main & wkup gpio intr
commit 3d54526e1abf5b80014068810a214d91ae7625c5 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The parent's input irq number is wrongly subtracted with 32 instead of
using the exact numbers in:

https://software-dl.ti.com/tisci/esd/latest/5_soc_doc/j721s2/interrupt_cfg.html

The GPIO interrupts are not working because of that. The toggling works
fine but interrupts are not firing. Fix the parent's input irq that
specifies the base for parent irq.

Tested for MAIN_GPIO0_6 interrupt on the j721s2 EVM.

Signed-off-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>