LF-5686: dmaengine: fsl-edma-v3: release descriptor before synchronize
commit 81986e1a1afddc8151560af395ef0268d7893bb4 from
https://github.com/nxp-imx/linux-imx.git

With the case that device_pause() is called, but device_terminate_all()
isn't called before device_synchronize(), the channel descriptor
is not released, which cause next instance can't device_issue_pending()
correctly.

Call the fsl_edma3_terminate_all() if the status is DMA_PAUSED in
device_synchronize() to handle this issue.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Joy Zou <joy.zou@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>