LF-5686: dmaengine: fsl-edma-v3: release descriptor before synchronize
commit ae871f43695bc0d86caab71f6218622aa2dc018f from
https://github.com/nxp-imx/linux-imx

With the case that device_pause() is called, but device_terminate_all()
isn't called before device_synchronize(), the channel descriptor
is not released, which cause next instance can't device_issue_pending()
correctly.

Call the fsl_edma3_terminate_all() if the status is DMA_PAUSED in
device_synchronize() to handle this issue.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Joy Zou <joy.zou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>