octeontx2-pcicons: Compute the padding size adaptively
For the SMP, the size of the arch_spinlock_t is always 4 bytes on
all architectures. So the change in the linux-yocto commit
5c8eed9bc9c1 ("octeontx2-pcicons: Adjust padding size for RT") makes
no sense, and would break the RT build due to the BUILD_BUG_ON()
sanity check in the pci_console_nexus_probe(). So we choose a more
adaptive way to compute the padding size so that it could work
even for UP.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>