LF-7107-1: media: capture: csi: Fix intermittent camera failures
commit be6dee3f8541f735497701fde50b32072bb55308 from
https://github.com/nxp-imx/linux-imx

When do stress test of capture, taking pictures and rebooting.
it will find that about 5% of them are black, the pictures shouldn't be black.

Root cause:
The few frames of ov5640 AE/AG is not stable, and at the begin of the few frames,
the ov5640 sensor 3A is under adjustment. And the picture data is dirty, so
it should be skipped at the start of the frame.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>