dwc: pci-s32cc: Use more than one MSI
commit bf7c8f82cd5b4d41122e3329faf33d31fc3a6039 from
https://github.com/nxp-auto-linux/linux

This patch gives PCIe the chance to use MSIs implemented as SPIs,
when configured as Root Complex and when "msi-parent" property
is defined in the PCIe node.
Until now the MSIs were generated using iMSI-RX and this mode
will still be used if "msi-parent" property is absent. The disadvantage
of this approach is that only one hardware interrupt can be used,
the one named "msi" from the interrupt list, which will make impossible
MSIs balancing between several cores.
If one wants to use more than one hardware interrupt, iMSI-RX should be
skipped and MSIs should be generated as SPIs.

From "PCIe controller databook 2020-05.pdf":
"3.8.2.3 iMSI-RX: Integrated MSI Receiver [AXI Bridge] The AXI
bridge provides an integrated MSI reception module to detect and
terminate inbound MSI requests (received on the RX wire). It is
enabled by setting CX_MSI_CTRL_ENABLE =1. Rather than propagating
MSI MWr TLPs onto the AXI bus through the master interface, the
MSI packets are captured, terminated in the AXI bridge, and an
interrupt is signaled locally through the msi_ctrl_int output.
The iMSI-RX is programmed with an address (MSI_CTRL_ADDR_OFF and
MSI_CTRL_UPPER_ADDR_OFF) that is used as the system MSI address.
When an inbound MWr request is passed to the AXI bridge and matches
this address as well as the conditions specified for an MSI memory
write request, an MSI interrupt is detected. When this MWr is about
to be driven onto the AXI bridge master interface1, it is dropped
and never appears on the AXI bus."

Issue: ALB-4796
Issue: ALB-6347
Issue: ALB-9108
Issue: ALB-10885
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>