MLK-19199 drm/imx: lcdif: replace FB width usage for cropping
commit ae8cd7b07ddbf5bb3510f859e303c957a9f1a1d3 from
https://github.com/nxp-imx/linux-imx.git

According to the comments of 'struct drm_framebuffer', its
'width' field refers to the logical width of the visible
area of the framebuffer. This may be unequal to the total
pixels number of a line. So use the 'pitches' field to
replace 'width' for the horizontal cropping feature.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
(cherry picked from commit 9a2bbbf971ed79b32ae1c7da2d62b8a72f3ccffd)
Signed-off-by: Meng Li <Meng.Li@windriver.com>