arm64: dts: ti: k3-am62x-lp-sk: disable PRU
commit c8d72cff5bda580f2f9621d4957773c3d11fe2fb from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Drop the PRU node from the devie tree on the AM62 LP SK boards, as
the AM62x SoC variants populated on those boards do not have one.

Attempting to access the PRU will result in a crash at kernel bootup.

At this time, we don't have any code that is able to identify the
absence of the PRU by reading the SoC registers. Do this in the board
DTS file for now, since all of the AMC package variants will not have
the PRU.

Signed-off-by: Anand Gadiyar <gadiyar@ti.com>
Cc: Andrew Davis <afd@ti.com>
Cc: Vignesh Raghavendra <vigneshr@ti.com>
Acked-by: Andrew Davis <afd@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>