LF-7194-3 arm64: dts: imx93-9x9-qsb: add spi-nor M.2 card support
commit ce758b9c530f4b1e6ce4fbf7c0f09850734e0a73 from
https://github.com/nxp-imx/linux-imx.git

There is a M.2 card has spi-nor device on it, so add this M.2 card support.

Currently, the M.2 card need to first enable the 3.3v voltage first, and
this 3.3v need about 1.74ms to change from 0v to 3.3v. Now u-boot already
enable this 3.3v, so Linux can ignore this 1.74ms delay. But if u-boot do
not do it, linux do need to take this 1.74ms into consideration, better
to build the flexspi driver as module to avoid this 1.74ms. Otherwise,
when read the spi-nor ID, the VCC of spi-nor device do not match the requirement.
always get 0xff, spi-nor will probe fail.

And the spi-nor RST pin is 1.1v default, this device require the RST pin in
1.8v, so need to change the register R4(on the M.2 board) from 10K to 1.5K.
If no this hardware rework, need to config the RST pin to logic high manually
through the pin12 of pcal6524 gpio expander.

Reviewed-by: Han Xu <han.xu@nxp.com>
Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>