i2c: xiic: Add wait for FIFO empty in send_tx
commit c5ce3375f5e8d046b06f9ac43a5aded419875901 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

If the tx_half_empty interrupt comes first instead of tx_empty,
STOP bit is generated even before all the bytes are transmitted
out on the bus.
STOP bit should be sent only after all the bytes in the FIFO are
transmitted out of the FIFO. So wait until FIFO is empty before sending
the STOP bit.

Signed-off-by: Raviteja Narayanam <raviteja.narayanam@xilinx.com>
Signed-off-by: Manikanta Guntupalli <manikanta.guntupalli@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
Link: https://lore.kernel.org/all/1663836294-5698-5-git-send-email-manikanta.guntupalli@xilinx.com
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>