ASoC: fsl_micfil: Add Hardware Voice Activity Detector support
commit 265d6e879ac8c0d25e6f82100ded1ebb33eae497 from
https://github.com/nxp-imx/linux-imx

The Hardware Voice Activity Detector (HWVAD) is a block
responsible for detect voice activity in a channel selected
by the user. It can be configured in Envelope-based or
Energy-based mode.

There are additional two interrupts for HWVAD, one is event
interrupt, another is error interrupt.

Enable hwvad in parallel with recording.
when voice activity detected, HWVAD will be disabled and
alsa control notification is triggerred.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>