LF-7261 clk: imx: keep hsio bus clock always on
commit 4e4c49e8d4cb62552f184eba8f09ea5bbb45a17c from
https://github.com/nxp-imx/linux-imx

When doing linux system suspend stress test after system sleep
enabled, sometimes, system will stuck. after some debug & narrow
down, it seems keep HSIO root clock ON can workaround this issue.

As the NICMIX power down can NOT be skipped for now, so it is hard
to check if this issue is related to the NICMIX power down.

so keep the HSIO root clock enabled as a temp workaround before we
have find out the root cause of the issue.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>