HSD #14017346372: hwmon: multiple location Temperature
commit c22da655dee11aa8173b8cfa246277d60b5bc8ac from
https://github.com/altera-opensource/linux-socfpga.git socfpga-5.15.100-lts.

The intend is to add support for multiple temperature sensing diodes
locations of Agilex die. TSDs are located in the core fabric and
transceiver tiles. The TSDs are read by SDM (Secure device manager)
through the HPS Mailbox to SDM. The change allows the device tree
<reg> to hold the location number in the upper byte of the 32bits
address and the TSD mask  number in the lower byte.
Add the missing device tree for hwmon on soc fpga with 3 default
TSDs in the Agilex Device tree for the SDM and HPS block.

Signed-off-by: Ang Tien Sung <tien.sung.ang@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>