LF-7294 i2c: imx-lpi2c: remove the redundant clk_bulk_prepare_enable calls
commit 2472afedc626cb5ffa9e82040b2f700a80093dea from
https://github.com/nxp-imx/linux-imx.git

Remove redundant clk_bulk_prepare_enable calls. pm_runtime_get_sync will be
called to open the clock. The call here will unbalance the clock switch,
causing the i2c clock to be occupied all the time.
This is a problem introduced when rebase.
This patch will be merged in code cleanup.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>