ethernet: gmac: Add necessary cleanup for GMAC RX clock
commit c9f00c0e3d73b1391a908393d572334c483172b3 from
https://github.com/nxp-auto-linux/linux

During unbind/bind command the GMAC RX clock has to be disabled.
For unbind of GMAC the source of the RX clock is the PHY which usually
is shutdown before the GMAC can do the necessary closing operations on
its RX clock. As a solution for unbind we closed the RX and TX GMAC
clks before the PHY.
A problem appears at bind also since the PHY is not UP at the probe of
GMAC driver. To solve this we enable the RX clock when setting the
TX clock speed since at that point the PHY is UP and the clock is
available.

Issue: ALB-8137
Signed-off-by: Andrei Botila <andrei.botila@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>