x86/resctrl: Create helper for RMID allocation and mondata dir creation
commit 4df9f713a16791779350a5d8c83fc6f7f2ffa2ba from
git@git.assembla.com:cavium/WindRiver.linux.git

RMID are allocated for each monitor or control group directory, because
each of these needs its own RMID. For control groups,
rdtgroup_mkdir_ctrl_mon() later goes on to allocate the CLOSID.

MPAM's equivalent of RMID are not an independent number, so can't be
allocated until the CLOSID is known. An RMID allocation for one CLOSID
may fail, whereas another may succeed depending on how many monitor
groups a control group has.

The RMID allocation needs to move to be after the CLOSID has been
allocated.

To make a subsequent change that does this easier to read, move the RMID
allocation and mondata dir creation to a helper.

Tested-by: Shaopeng Tan <tan.shaopeng@fujitsu.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>