octeontx2-pcicons: provide toggle for trace_printk
commit f68568783588cfb919858b8ff071e3edf08c0bfe from
git@git.assembla.com:cavium/WindRiver.linux.git

If kernel is built with tracing capabilities, presence of trace_printk()
call anywhere in kernel code causes kernel to allocate extra buffers for
tracing purposes (on boot time, even if tracing is disabled). A warning
is shown in kernel log as well.

Provide a compile-time switch for tracing enablement in the PCIe console
driver, so it can be enabled during developer build manually as needed.

Change-Id: Ic3cca188bbc1a1a36846db41ce460cb445bac635
Signed-off-by: Wojciech Zmuda <wzmuda@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/88070
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>