media: dt-bindings: Convert imx290.txt to YAML
commit  cc28b47d409b22f039ec1a57a68ee23af6716953 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

Upstream commit a96dfea1df25

Convert the Sony IMX290 DT binding from text to YAML. Add Manivannan as
a maintainer given that he is listed in MAINTAINERS for the file, as
volunteering myself.

The name of the input clock, "xclk", is wrong as the hardware manual
names it INCK. As the device has a single clock, the name could be
omitted, but that would require a corresponding change to the driver and
is thus a candidate for further patches.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>