media: i2c: imx290: Replace macro with explicit ARRAY_SIZE()
commit  c5184fd0ba37f5267042cd8ecaf7d7032be52fbc from
https://github.com/raspberrypi/linux.git rpi-6.1.y

Upstream commit 2548df538cdd.

Use ARRAY_SIZE(imx290->supplies) for code that needs the size of the
array, instead of relying on the IMX290_NUM_SUPPLIES. The result is less
error-prone as it ties the size to the array.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Alexander Stein <alexander.stein@ew.tq-group.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>