media: i2c: imx290: Drop imx290_write_buffered_reg()
commit  93944049a369608a69d6815bc40e35cb27cefb14 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

Upstream commit b817888a0c50.

The imx290_write_buffered_reg() function wraps a register write with
register hold, to enable changing multiple registers synchronously. It
is used for the gain only, which is an 8-bit register, defeating its
purpose.

The feature is useful, but should be implemented differently. Drop the
function for now, to prepare for a rework of register access.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Acked-by: Alexander Stein <alexander.stein@ew.tq-group.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>