octeontx2-bphy-netdev: Add new secondary BCN offset for slave
commit 32a82be4f7ff6759e1558e63b43d1fb73b456fff from
git@git.assembla.com:cavium/WindRiver.linux.git

The secondary BCN offset is different for master and slave, so
using a single parameter in common for master and slave is incorrect.
Hence a new secondary BCN offset parameter is added to "netdev_priv"
structure and is used to sychronize BCN to PTP clock in slave mode. This
parameter can be updated using the IOCTL "BCN_PTP_SYNC".
In addition to this a new parameter is added to structure "bcn_ptp_cfg",
which holds the offset between SW timecounter and ToD and is returned to
the user on calling the IOCTL "BCN_PTP_SYNC".

Change-Id: I00c7edffd0d64ffc6e4345598514259bb6ccc561
Signed-off-by: Rakesh Babu Saladi <rsaladi2@marvell.com>
Fixes: b0d0d83ebe95("octeontx2-bphy-netdev: PTP BCN synchronization
		     support for CNF10k platforms.")
Fixes: 635f93a93887("otx2-bphy-netdev: synchronize BCN to PTP slave
		     clock.")
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/88991
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Naveen Mamindlapalli <naveenm@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>