irqchip/gic-v3: Add workaround for interrupt loss on IPI
commit 2d47e6e8d19cc489da56ba46f07a69a53a136269 from
git@git.assembla.com:cavium/WindRiver.linux.git

RCU stalls has been observed during stress-ng tests on
OcteonTx2 silicon and the reason being an interrupt loss on an IPI.

As a workaround, added provision to ensure that every IPI has been
received on the destination core and if an IPI interrupt loss is
encountered, do resend the IPI.

Identifying a lost interrupt:
Once an interrupt is raised on an IPI line, it would be either in
one of the below states.
   A. pending
   B. active
   C. deactivated

Pending/active states can be verified from the corresponding redist
registers. An interrupt would be in deactivated state either because its
already processed or its been lost. So to identify the reason for case C,
per core tx and rx counters has been introduced for each IPI line.

Change-Id: I02c0b885cbdef8084d9a08aebaf3e6bba33f97e2
Signed-off-by: Linu Cherian <lcherian@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/89369
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>