octeontx2-bphy-netdev:cnf10k: Share RFOE PTP offset with host
commit 5d659581f68c440d005a7eeee098bdf00aac0b12 from
git@git.assembla.com:cavium/WindRiver.linux.git

Share the offset that is added to the PTP counter to get the
UTC time with the host.  This enables the host to use this offset
along with reading the PTP counter to directly use the Octeon EP
PTP counter as a timesource.  The offset is stored in
the unused MIO_PTP_CKOUT_THRESH_HI register.
Note that this offset is not directly maintained in the driver,
but calculated at initialization time, then updated as needed.
This is based on the same feature previously added to cnf95n

Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: Ib913868ea36265f6e6a4b0efb1d5e24bb89cbb90
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/89664
Reviewed-by: Satananda Burla <sburla@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>