LF-7486-1 edac: fsl_ddr: Add support for i.MX9 DDR controller
commit 6e530219610ffc1fe10d0ddf9fb990611b1e1175 from
https://github.com/nxp-imx/linux-imx.git

i.MX9 uses an DDR controller that has registers offset and some
function changed, but the ECC and Error inject functions are
almost same with fsl_ddr controller. So update and re-use the
driver for i.MX9. A special type is added for i.MX9 controller
only to distinguish the difference.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>