LF-7486-1 edac: fsl_ddr: Add support for i.MX9 DDR controller
commit 4b0cb3019697cb3cbbf9d6be69008f9300d708bb from
https://github.com/nxp-imx/linux-imx

i.MX9 uses an DDR controller that has registers offset and some
function changed, but the ECC and Error inject functions are
almost same with fsl_ddr controller. So update and re-use the
driver for i.MX9. A special type is added for i.MX9 controller
only to distinguish the difference.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>