LF-7981-4: drm/bridge: nwl-dsi: Correct MIPI DSI horizontal sync timing
commit c01372b8cf85039cc929b1aef4c94868ff4dc933 from
https://github.com/nxp-imx/linux-imx.git

The NWL MIPI Host controller registers set the number of bytes for the
horzontal front porch, sync pulse, and back porch, not the number of
pixels. The formula converts the hfp, hsa, and hbp to bytes then subtracts
the number of packet overhead bytes in the horizontal line which totals 32.
The overhead is split into three proportional chunks and subtracted from
fp, hsa, and hbp.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Oliver F. Brown <oliver.brown@oss.nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>