net: fec: simplify the code logic of quirks
commit  35d0b61d4c84247e35cb402b40519b97dccbf369 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

Simplify the code logic of handling the quirk of FEC_QUIRK_HAS_RACC.
If a SoC has the RACC quirk, the driver will enable the 16bit shift
by default in the probe function for a better performance.

This patch handles the logic in one place to make the logic simple
and clean. The patch optimizes the fec_enet_xdp_get_tx_queue function
according to Paolo Abeni's comments, and it also exludes the SoCs that
require to do frame swap from XDP support.

Signed-off-by: Shenwei Wang <shenwei.wang@nxp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
[Xiaolei: This is only different from the original patch
in terms of context, but not in content.]
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>