drm/vc4: crtc: Make encoder lookup helper public
commit  a1bf398426a27fb201029944c08faf656a98d8d9 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

We'll need a function that looks up an encoder by its vc4_encoder_type.
Such a function is already present in the CRTC code, so let's make it
public so that we can reuse it in the unit tests.

Reviewed-by: Javier Martinez Canillas <javierm@redhat.com>
Reviewed-by: Maíra Canal <mcanal@igalia.com>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>