drm/vc4: hdmi: Add property to allow manual config of RGB or YCbCr
commit  605a5ae7bbf68ed579015c3a5bb698d89723a369 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

Add a custom property "Output format" that allows the overriding
of the default colourspace choice in the way that the old
firmware hdmi_pixel_encoding property did. If the chosen format is not
supported, then it will still drop back to the older behaviour.

This won't be acceptable to upstream, but it adds back the missing
functionality of hdmi_pixel_encoding.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>