edac: deprecate the cortex cache edac driver
commit 4045fa99a7ca045db856d79eb96417ad5a9b1d56 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

The current edac driver is violating a arm recommendation
of accessing the actlr register when not idle.

The accessibility of the actlr is dependent on
the ACTLR.EL3 and ACTLR.EL2 so it may crash some
implementations.

The driver is rejected in mainline [1] as it does not support
the other implementatons like big.LITTLE. We are deprecating
the driver. It may be deleted in later in 2023.3.

[1] https://lore.kernel.org/all/20151028164053.GG25451@leverpostej/

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>