net: ethernet: ti: davinci_mdio: Deferring probe when soc_device_match() returns NULL
commit d17c68db2a164beaeec8d0acd5ea78a79f69161d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When the k3 socinfo driver is built as a module, there is a possibility
that it will probe after the davinci mdio driver. By deferring the mdio
probe we allow the k3 socinfo to probe and register the
soc_device_attribute structure needed by the mdio driver.

Signed-off-by: Nicolas Frayer <nfrayer@baylibre.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>